requirements.txt
async-generator==1.10
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
exceptiongroup==1.1.1
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-oauth==1.0.1
gspread==5.7.2
h11==0.14.0
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
webdriver-manager==3.8.5
wsproto==1.2.0
